fix: refactor signup logic
seperate business logic into service file and request and response
into controller file.
